Separate edit and exporting
