+++ visualizer/src/Project/DisplayControls/ExportButton/SubmitButton.js
@@ -3,19 +3,18 @@ import { Button, CircularProgress } from '@mui/material';
 import { green } from '@mui/material/colors';
 import { useSelector } from '@xstate/react';
 import React from 'react';
-import { useApi } from '../../ProjectContext';
+import { useExport } from '../../ProjectContext';
 
 function SubmitButton() {
-  const api = useApi();
-  const { send } = api;
-  const uploading = useSelector(api, (state) => state.matches('uploading'));
+  const export_ = useExport();
+  const uploading = useSelector(export_, (state) => state.matches('uploading'));
 
   return (
     <Button
       variant='contained'
       color='primary'
       endIcon={<SendIcon />}
-      onClick={() => send('UPLOAD')}
+      onClick={() => export_.send('UPLOAD')}
       disabled={uploading}
       sx={{ position: 'relative' }}
     >
